rename "credentials" folder to "dynamic"
