+++ src/dynamic.ts
@@ -1,23 +1,20 @@
-export { DynamicArray } from './credentials/dynamic-array.ts';
-export { StaticArray } from './credentials/static-array.ts';
-export { DynamicBytes } from './credentials/dynamic-bytes.ts';
-export { DynamicString } from './credentials/dynamic-string.ts';
-export { DynamicRecord } from './credentials/dynamic-record.ts';
+export { DynamicArray } from './dynamic/dynamic-array.ts';
+export { StaticArray } from './dynamic/static-array.ts';
+export { DynamicBytes } from './dynamic/dynamic-bytes.ts';
+export { DynamicString } from './dynamic/dynamic-string.ts';
+export { DynamicRecord } from './dynamic/dynamic-record.ts';
 export {
   DynamicSHA2,
   Sha2IterationState,
   Sha2Iteration,
   Sha2FinalIteration,
-} from './credentials/dynamic-sha2.ts';
-export { DynamicSHA3 } from './credentials/dynamic-sha3.ts';
+} from './dynamic/dynamic-sha2.ts';
+export { DynamicSHA3 } from './dynamic/dynamic-sha3.ts';
 export {
   toDecimalString,
   toBaseBE,
   fromBaseBE,
-} from './credentials/gadgets-digits.ts';
+} from './dynamic/gadgets-digits.ts';
 
-export {
-  hashDynamic,
-  hashDynamicWithPrefix,
-} from './credentials/dynamic-hash.ts';
-export { Schema } from './credentials/schema.ts';
+export { hashDynamic, hashDynamicWithPrefix } from './dynamic/dynamic-hash.ts';
+export { Schema } from './dynamic/schema.ts';
