+++ src/dynamic/dynamic-bytes.ts
