+++ src/dynamic/dynamic-record.ts
