+++ src/dynamic/dynamic-string.ts
