+++ src/dynamic/keccak-permutation.ts
