+++ src/dynamic/static-array.ts
