+++ src/index.ts
@@ -10,17 +10,17 @@ export {
   ZkAppRequest,
 } from './presentation.ts';
 export { assert } from './util.ts';
-export { DynamicArray } from './credentials/dynamic-array.ts';
-export { StaticArray } from './credentials/static-array.ts';
-export { DynamicBytes } from './credentials/dynamic-bytes.ts';
-export { DynamicString } from './credentials/dynamic-string.ts';
-export { DynamicRecord } from './credentials/dynamic-record.ts';
-export { DynamicSHA2 } from './credentials/dynamic-sha2.ts';
+export { DynamicArray } from './dynamic/dynamic-array.ts';
+export { StaticArray } from './dynamic/static-array.ts';
+export { DynamicBytes } from './dynamic/dynamic-bytes.ts';
+export { DynamicString } from './dynamic/dynamic-string.ts';
+export { DynamicRecord } from './dynamic/dynamic-record.ts';
+export { DynamicSHA2 } from './dynamic/dynamic-sha2.ts';
 export { hashPacked } from './o1js-missing.ts';
 export {
   hashDynamic,
   hashDynamicWithPrefix,
   log,
   toValue,
-} from './credentials/dynamic-hash.ts';
-export { Schema } from './credentials/schema.ts';
+} from './dynamic/dynamic-hash.ts';
+export { Schema } from './dynamic/schema.ts';
